+++ src/converters/degrees2coordinates_converter.py
@@ -1,15 +1,8 @@
-from typing import List
-
-import pandas as pd
 from pandas import DataFrame
-import re
-import numpy as np
 
-from src.converters.utility import format_degrees_to_coordinate_lat, format_degrees_to_coordinate_long, \
-    convert_lat_long_cols
-from src.exceptions import ConversionFailedException
+from src.converters.utility import convert_lat_long_cols
 from src.interfaces.converter_interface import IConverter
-from src.utility.utility import get_exception
+from src.utility import get_exception
 
 from src.log import get_logger
